@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,18 +27,27 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  flutter_spinkit: ^4.1.0
-  http: ^0.12.2
-  url_launcher: ^5.7.10
-  intl: ^0.16.1
-  flutter_html: ^1.2.0
+  cupertino_icons: ^1.0.3
+  flutter_spinkit: ^5.0.0
+  http: ^0.13.3
+  url_launcher: ^6.0.9
+  intl: ^0.17.0
+  flutter_html: ^2.1.0
+
+  firebase_core: ^1.3.0
+  firebase_messaging: ^10.0.3
+  flutter_local_notifications: ^6.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   integration_test:
     sdk: flutter
+
+dependency_overrides:
+  archive: ^3.0.0
+  convert: ^3.0.0
+  crypto: ^3.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
