@@ -9,7 +9,112 @@ import 'package:infinite_mobile_app/pages/mission.dart';
 import 'package:infinite_mobile_app/pages/error_page.dart';
 import 'package:infinite_mobile_app/pages/event_page.dart';
 
-void main() => runApp(MaterialApp(initialRoute: '/loading', routes: {
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+
+const AndroidNotificationChannel channel = AndroidNotificationChannel(
+    'high_importance_channel', // id
+    'High Importance Notifications', // title
+    'This channel is used for important notifications.', // description
+    importance: Importance.high,
+    playSound: true);
+
+final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
+    FlutterLocalNotificationsPlugin();
+
+Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
+  await Firebase.initializeApp();
+  print('A bkg message just showed up :  ${message.messageId}');
+}
+
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+  FirebaseMessaging.onBackgroundMessage(_firebaseMessagingBackgroundHandler);
+
+  await flutterLocalNotificationsPlugin
+      .resolvePlatformSpecificImplementation<
+          AndroidFlutterLocalNotificationsPlugin>()
+      ?.createNotificationChannel(channel);
+
+  await FirebaseMessaging.instance.setForegroundNotificationPresentationOptions(
+    alert: true,
+    badge: true,
+    sound: true,
+  );
+
+  runApp(InitializationRoutines());
+}
+
+class InitializationRoutines extends StatefulWidget {
+  @override
+  _InitRoutesAndNotifications createState() => _InitRoutesAndNotifications();
+}
+
+String ReturnNonNullableString(String? test_string) {
+  if (test_string == null) {
+    test_string = "";
+  }
+  return test_string;
+}
+
+class _InitRoutesAndNotifications extends State<InitializationRoutines> {
+  // @override
+  // void initState() {
+  //   super.initState();
+  // FirebaseMessaging.onMessage.listen((RemoteMessage message) {
+  //   RemoteNotification? notification = message.notification;
+  //   AndroidNotification? android = message.notification?.android;
+  //   if (notification != null && android != null) {
+  //     flutterLocalNotificationsPlugin.show(
+  //         notification.hashCode,
+  //         notification.title,
+  //         notification.body,
+  //         NotificationDetails(
+  //           android: AndroidNotificationDetails(
+  //             channel.id,
+  //             channel.name,
+  //             channel.description,
+  //             color: Colors.blue,
+  //             playSound: true,
+  //             icon: '@mipmap/ic_launcher',
+  //           ),
+  //         ));
+  //   }
+  // }
+  // );
+
+  // FirebaseMessaging.onMessageOpenedApp.listen((RemoteMessage message) {
+  //   print('A new onMessageOpenedApp event was published!');
+
+  //   RemoteNotification? notification = message.notification;
+  //   AndroidNotification? android = message.notification?.android;
+
+  //   String NonNullableTitle = ReturnNonNullableString(notification.title);
+  //   // Argggggg!!!! This doesn't work
+
+  //   if (notification != null && android != null) {
+  //     showDialog(
+  //         context: context,
+  //         builder: (_) {
+  //           return AlertDialog(
+  //             title: Text(notification.title),
+  //             content: SingleChildScrollView(
+  //               child: Column(
+  //                 crossAxisAlignment: CrossAxisAlignment.start,
+  //                 children: [Text(notification.body)],
+  //               ),
+  //             ),
+  //           );
+  //         });
+  //   }
+  // });
+  // }
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(initialRoute: '/loading', routes: {
       '/loading': (context) => Loading(),
       '/home': (context) => Home(),
       '/contact': (context) => Contact(),
@@ -18,4 +123,6 @@ void main() => runApp(MaterialApp(initialRoute: '/loading', routes: {
       '/mission': (context) => Mission(),
       '/event_page': (context) => EventPage(),
       '/error_page': (context) => ErrorPage(),
-    }));
+    });
+  }
+}
