@@ -9,7 +9,7 @@ class Donate extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: InfiniteAppBar(context),
+        appBar: InfiniteAppBar(context) as PreferredSizeWidget?,
         body: SingleChildScrollView(
           padding: const EdgeInsets.only(top: 20.0),
           child: Column(children: <Widget>[
